MCPJVM-344: Move utils for performance-suite to its own dedicated /spec folder
