+++ docs/specs/README.md
@@ -4,6 +4,13 @@ This folder contains spec-style contracts that are intended to be both human-rea
 
 Current specs:
 
+- `execution-plan-spec`
+  - suite-neutral execution-plan models and path policy
+- `performance-execution-plan-spec`
+  - performance execution-plan models and workload-provider contract
 - `regression-execution-plan-spec`
-  - `performance-msta-evidence-model.md`: normative MSTA evidence contract for performance-suite timing analysis
+  - regression execution-plan models and validation contracts
 
+Performance evidence documentation remains under the regression spec documentation tree for compatibility:
+
+- `regression-execution-plan-spec/performance-msta-evidence-model.md`: normative MSTA evidence contract for performance-suite timing analysis
