+++ tools/features/execution-orchestration/shared/resiliency.ts
@@ -1,7 +1,7 @@
 import type {
   RuntimeSuiteBlockedResult,
   RuntimeSuiteRunResult,
-} from "../../../spec/regression-execution-plan-spec/src/models/regression_runtime_suite.model";
+} from "@tools-execution-plan-spec/models/runtime_suite.model";
 import type {
   ExecutionOrchestrationLoopDefaults,
   ExecutionOrchestrationLoopPolicy,
