+++ tools/features/execution-orchestration/support/execution_orchestration_suite.ts
@@ -3,7 +3,7 @@ import { dispatchRegressionSuiteAction } from "@tools-regression-suite";
 import type {
   RuntimeSuiteBlockedResult,
   RuntimeSuiteRunResult,
-} from "../../../spec/regression-execution-plan-spec/src/models/regression_runtime_suite.model";
+} from "@tools-execution-plan-spec/models/runtime_suite.model";
 import type {
   ExecutionOrchestrationPassState,
   ExecutionOrchestrationPassResult,
