+++ tools/features/execution-profile-export/loaders/export_target.loader.ts
@@ -2,7 +2,7 @@ import { promises as fs } from "node:fs";
 import path from "node:path";
 
 import { readProjectArtifact } from "@tools-feature-artifact-management";
-import type { ExecutionProfileSuiteType } from "../../../spec/regression-execution-plan-spec/src/models/regression_execution_profile_export.model";
+import type { ExecutionProfileSuiteType } from "@tools-execution-plan-spec/models/execution_profile_export.model";
 
 import { asString, isRecord } from "../common";
