+++ tools/spec/execution-plan-spec/src/models/suite_run_status.model.ts
@@ -1,0 +1,16 @@
+export type SuiteRunStatus = "pass" | "fail" | "blocked" | "in_progress";
+
+export type SuiteWatcherPhaseStatus =
+  | "not_configured"
+  | "pass"
+  | "fail"
+  | "blocked"
+  | "in_progress";
+
+export type SuiteExternalVerificationPhaseStatus =
+  | "not_configured"
+  | "pass"
+  | "fail"
+  | "blocked"
+  | "in_progress"
+  | "skipped_dependency";
